# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bhulan GPS API",
    "description": "Mobility insights and map plotting for GPS coordinates, plus ingestion endpoints for vendor feeds.",
    "version": "2.1.0"
  },
  "paths": {
...
    "/ingest/trackpoints": {
      "post": {
        "summary": "Ingest Trackpoints",
        "description": "Ingest GPS track points via webhook.\n\nAccepts JSON payload with GPS data, normalizes it, and persists to database.\n\nArgs:\n    payload: JSON object or array of GPS records\n    vendor: Vendor identifier (generic, geotab, samsara)\n    ingest_id: Optional ingestion job ID (generated if not provided)\n    x_bhulan_mapping: Optional custom mapping JSON in header\n\nReturns:\n    NormalizationResult with accepted/rejected counts and errors",
        "operationId": "ingest_trackpoints_ingest_trackpoints_post",
        "parameters": [
          {
            "name": "vendor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Vendor/source identifier",
              "default": "generic",
              "title": "Vendor"
            },
            "description": "Vendor/source identifier"
          },
          {
            "name": "ingest_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Ingestion job ID",
              "title": "Ingest Id"
            },
            "description": "Ingestion job ID"
          },
          {
            "name": "x-bhulan-mapping",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Custom mapping JSON",
              "title": "X-Bhulan-Mapping"
            },
            "description": "Custom mapping JSON"
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "additionalProperties": true
                    }
                  },
                  {
                    "type": "object",
                    "additionalProperties": true
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NormalizationResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{ingest_id}": {
      "get": {
        "summary": "Get Job Status",
        "description": "Get ingestion job status and statistics.\n\nArgs:\n    ingest_id: Ingestion job identifier\n\nReturns:\n    Job information including status, stats, and errors",
        "operationId": "get_job_status_jobs__ingest_id__get",
        "parameters": [
          {
            "name": "ingest_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ingest Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NormalizationResult": {
        "properties": {
          "accepted": {
            "type": "integer",
            "title": "Accepted",
            "description": "Number of records accepted"
          },
          "rejected": {
            "type": "integer",
            "title": "Rejected",
            "description": "Number of records rejected"
          },
          "errors": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Errors",
            "description": "Row index to error message mapping"
          },
          "ingest_id": {
            "type": "string",
            "title": "Ingest Id",
            "description": "Ingestion job ID"
          }
        },
        "type": "object",
        "required": [
          "accepted",
          "rejected",
          "ingest_id"
        ],
        "title": "NormalizationResult",
        "description": "Result of normalizing a batch of GPS data."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}